Fixed an issue which caused CPU metric to be missing in azurewebapp due to overflow of int variable used to read cpu times.

+++ Src/PerformanceCollector/Shared/Implementation/WebAppPerformanceCollector/ICachedEnvironmentVariableAccess.cs
@@ -11,6 +11,6 @@
         /// <param name="name"> Name of the counter.</param>
         /// <param name="environmentVariable"> Identifier of the corresponding environment variable.</param>
         /// <returns> Counter value.</returns>
-        int GetCounterValue(string name, AzureWebApEnvironmentVariables environmentVariable);
+        double GetCounterValue(string name, AzureWebApEnvironmentVariables environmentVariable);
     }
 }
